+++ spec/lib/secure_headers/configuration_spec.rb
@@ -5,83 +5,33 @@ module SecureHeaders
   describe Configuration do
     before(:each) do
       reset_config
-      Configuration.default
     end
 
     it "has a default config" do
-      expect(Configuration.get(Configuration::DEFAULT_CONFIG)).to_not be_nil
+      expect(Configuration.default).to_not be_nil
     end
 
-    it "has an 'noop' config" do
-      expect(Configuration.get(Configuration::NOOP_CONFIGURATION)).to_not be_nil
+    it "has an 'noop' override" do
+      Configuration.default
+      expect(Configuration.overrides(Configuration::NOOP_OVERRIDE)).to_not be_nil
     end
 
-    it "precomputes headers upon creation" do
-      default_config = Configuration.get(Configuration::DEFAULT_CONFIG)
-      header_hash = default_config.cached_headers.each_with_object({}) do |(key, value), hash|
-        header_name, header_value = if key == :csp
-          value["Chrome"]
-        else
-          value
-        end
-
-        hash[header_name] = header_value
-      end
-      expect_default_values(header_hash)
-    end
-
-    it "copies config values when duping" do
-      Configuration.override(:test_override, Configuration::NOOP_CONFIGURATION) do
-        # do nothing, just copy it
-      end
-
-      config = Configuration.get(:test_override)
-      noop = Configuration.get(Configuration::NOOP_CONFIGURATION)
-      [:csp, :csp_report_only, :cookies].each do |key|
-        expect(config.send(key)).to eq(noop.send(key))
+    it "dup results in a copy of the default config" do
+      Configuration.default
+      original_configuration = Configuration.send(:default_config)
+      configuration = Configuration.dup
+      expect(original_configuration).not_to be(configuration)
+      Configuration::CONFIG_ATTRIBUTES.each do |attr|
+        expect(original_configuration.send(attr)).to eq(configuration.send(attr))
       end
     end
 
-    it "regenerates cached headers when building an override" do
-      Configuration.override(:test_override) do |config|
-        config.x_content_type_options = OPT_OUT
-      end
-
-      expect(Configuration.get.cached_headers).to_not eq(Configuration.get(:test_override).cached_headers)
-    end
-
-    it "stores an override of the global config" do
+    it "stores an override" do
       Configuration.override(:test_override) do |config|
         config.x_frame_options = "DENY"
       end
 
-      expect(Configuration.get(:test_override)).to_not be_nil
-    end
-
-    it "deep dup's config values when overriding so the original cannot be modified" do
-      Configuration.override(:override) do |config|
-        config.csp[:default_src] << "'self'"
-      end
-
-      default = Configuration.get
-      override = Configuration.get(:override)
-
-      expect(override.csp.directive_value(:default_src)).not_to be(default.csp.directive_value(:default_src))
-    end
-
-    it "allows you to override an override" do
-      Configuration.override(:override) do |config|
-        config.csp = { default_src: %w('self'), script_src: %w('self')}
-      end
-
-      Configuration.override(:second_override, :override) do |config|
-        config.csp = config.csp.merge(script_src: %w(example.org))
-      end
-
-      original_override = Configuration.get(:override)
-      expect(original_override.csp.to_h).to eq(default_src: %w('self'), script_src: %w('self'))
-      override_config = Configuration.get(:second_override)
-      expect(override_config.csp.to_h).to eq(default_src: %w('self'), script_src: %w('self' example.org))
+      expect(Configuration.overrides(:test_override)).to_not be_nil
     end
 
     it "deprecates the secure_cookies configuration" do
@@ -101,7 +51,7 @@ module SecureHeaders
         config.cookies = OPT_OUT
       end
 
-      config = Configuration.get
+      config = Configuration.dup
       expect(config.cookies).to eq(OPT_OUT)
     end
 
@@ -110,7 +60,7 @@ module SecureHeaders
         config.cookies = {httponly: true, secure: true, samesite: {lax: false}}
       end
 
-      config = Configuration.get
+      config = Configuration.dup
       expect(config.cookies).to eq({httponly: true, secure: true, samesite: {lax: false}})
     end
   end